backend/requirements.txt
# flask==3.0.0
# flask-cors==4.0.0
# mysql-connector-python==8.2.0
# openai==1.6.0
# python-dotenv==1.0.0
# sentence-transformers==2.2.2
# faiss-cpu==1.7.4
# numpy==1.24.3
# gunicorn==21.2.0
flask==3.0.0
flask-cors==4.0.0
mysql-connector-python==8.2.0
openai>=1.30.0
python-dotenv==1.0.0
sentence-transformers>=2.3.0
faiss-cpu==1.7.4
numpy>=1.23.0,<2.0
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.3.0
scipy>=1.10.0
huggingface-hub>=0.17.0
gunicorn==21.2.0